pyproject.toml
[project]
name = "openmux"
version = "0.1.0"
description = "A free multi-source GenAI orchestration library for automatic model selection and routing"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["ai", "llm", "genai", "model-selection", "orchestration", "free-models", "openmux"]
authors = [
    {name = "Nizam Uddin", email = "your.email@example.com"}
]
maintainers = [
    {name = "Nizam Uddin", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0",
    "keyring>=24.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "types-requests",
]

[project.urls]
Homepage = "https://github.com/mdnu838/OpenMux"
Documentation = "https://github.com/mdnu838/OpenMux/tree/main/docs"
Repository = "https://github.com/mdnu838/OpenMux"
Issues = "https://github.com/mdnu838/OpenMux/issues"
Changelog = "https://github.com/mdnu838/OpenMux/blob/main/CHANGELOG.md"

[project.scripts]
openmux = "openmux.cli.main:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["openmux"]

[tool.setuptools.package-data]
openmux = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
